+++ src/main/java/view/View.java
@@ -1,4 +1,4 @@
-package controller;
+package view;
 
 import http.response.HttpResponse;
