+++ src/main/java/view/ViewResolver.java
@@ -1,4 +1,4 @@
-package controller;
+package view;
 
 import java.util.HashMap;
 import java.util.Map;
